BugFixes and enhancements
Lowered the wave surface level to fix an issue where sometimes double jump is disabled.

Enhancements - We realised that there was a siginificant drop in FPS because up to three patterns could exist, and that means too many instantiated objects.

To fix this:
Included a second destroyer - which only destroys Obstacles, Coins and Waves.

Pushed window towards the end of the pattern so that new pattern generation is delayed.